refactor: Move ZIO module to zio subpackage
